Move MonadFileSystem into its own (private) library.
I am currently reflecting on the fact that I decided to write an
in-memory filesystem just to speed up some tests. It occurs to me that
this is complex enough to probably merit its own tests.

One of the following will have happen:

1. I will find time to spin of MonadFileSystem as its own project, get
   it working and polish it.
2. I will make TEST_IN_IO the default for ddb, and avoid doing any more
   yak shaving for the time being.

(2) Will probably happen first, with (1) maybe being a future project
down the road.