Check the error type when loading metadata fails.
We should only fall-back if it's because the file was absent.

This doesn't seem to have an observable effect, so not responsible for
the failures.

Playing around at the repl, it appears that the problem *is* in decoding
the metadata, but the error doesn't make sense to me; something about
expecting 6 elements but getting 3. Inspecting the file confirms that
the top-level object is a 3 element list. But we've derived our encode &
decode, so I'm not sure why they don't match up; will have to
investigate further.