More useful test output.
It actually shows us a diff now (along with some info about it).

This pointed me at what I think is the source of the problem: we're
truncating the blob file on subsequent opens. Shouldn't be too hard to
fix.

The CBOR error was a red herring -- just me fat-fingering and typing
Metadata (from DDB.Types) instead of MetaData (from NewStore). I ought
to rename one of those...