Don't overwrite the old blob file.
This was the cause of the test failures.

I'd added this before because I thought openBinaryFile was expected a
pre-existing file, but the error was actually about the store directory
not existing, rather than the blob file. Doing the
createDirectoryIfMissing for the tags fixes that by itself, so we don't
need this.

A first manual test of the tool seems to suggest it is working as well,
but we should do more testing.